requirements.txt
# Minimal requirements for the notebook
torch>=2.1
torchvision>=0.16
numpy>=1.23
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
pillow>=10.0
tqdm>=4.66
